Finished the KMeans Algorithm
Finished writing the document for the KMeans project involving a data set of segmented frog calls prelabeled based on family, genus, etc. Interesting discovery was that the K-Means algorithm discovered that the ideal number of clusters was 4 and yet the it is not representative of the family of frogs within the data set.

Some more research regarding the intrisic properties of the clusters may be due.